Add support for regulated fields
Allow the content rules to relate to policies in a non-disruptive way.
Some policies may be very picky when it comes to implementation of
configuration requirements - the wording can be prescribed in a literal
way that could interfere with the otherwise neutral and general
description/rationale.

This change allows definition of such rule fields in separate files
in a free-form schema.