Clarify instructions for implementing SCCs
We advise users to deploy security context constraints to secure pods.
This is especially important with OpenShift 4, which comes with default
security context constraints enabled and ready to use.

Previously, the documentation for using security context constraints was
vague. This is somewhat understandable since using security context
constrains is open-ended, and really depends on application and pods
running within a cluster.

This commit tries to clarify the instructions by briefly introducing
security context constraints, and leaving the educational heavy lifting
to another OpenShift document that's dedicated to describing that
feature and how to use it effectively.

Jira: https://issues.redhat.com/browse/OCPBUGS-698