Use of variables in calculations and PEP8 fixes
In the create_implicit_control_lists function lists are used multiple
times and it is better to defined local variables for them in favor of
readability. This was also applied to the count_implicit_status
function. Although this last function does not repeat values, the
variables are still good for readability.