Recreated OVAL assessment for firewalld_sshd_port_enabled
There are some corner cases involving possible realistic scenarios with
firewalld and NetworkManager. Based on the remediation refactoring, the
OVAL assessment was also reformulated to be more simple and much more
reliable. It is now checking firewalld packaged files and also custom
files respecting the proper order in case of custom files.