Increase robustness of firewalld_sshd_port_enabled bash remediation
The remediation was not capable to properly treat some special cases,
like a system with multiple interfaces. It wasn't also capable to safely
configure the correct interface since it was assuming the NetworkManager
connection file was prefixed with the network interface name. In
addition, it is not stable to manually change firewalld XML files while
a proper command is present. This commit makes the remediation reliable
and assertive by using firewall-cmd and nmcli commands.