Reinvent the test scenarios for firewalld_sshd_port_enabled
The test scenarios were aligned to the old remediation approach, making
them also incomplete and incapable to catch real cases. Once the
remediation was robust, test scenarios also need the same level of
robustness in order to ensure the rules is as much realistic as
possible. They are now covering cases with multiple interfaces and
multiple active zones. It is also covered custom SSH port.