Use diff_files instead of diff_trees
Originally, we thought xmldiff is sensitive on whitespace before and
after tags. But now we have found that this isn't true for the
diff_files method. When using the diff_trees it is sensitive on
whitespace and when using diff_files method it isn't sensitive on
whitespace.

So we can save the actual outputs produced by the SUT to temporary files
and then compare the temporary files with the test data files using the
diff_files.