Create accounts_passwords_pam_faillock_audit rule
This rule was based on a existing rule with similar name:
account_passwords_pam_faillock_audit
The existing rule name was not aligned to the other similar rules making
it prone to confusion. This commit duplicates code but it is worth in
order to improve the project consistency. This will be solved in the
future by deprecating the rule with inconsistent name.