Increment remediation to ensure correct SELinux fcontext
This remediation was setting the permanent dir for tally logs, but was
not ensuring this new directory was properly configured with the
faillog_t SELinux context. As a consequence, in a second assessment of
relevant profiles would fail in the account_password_selinux_faillock_dir
rule. The Ansible and Bash remediation were updated to ensure the proper
SELinux file context.