OCP4: lowercase several rules
The OCP4 e2e test suite matches the rules coming from the kubernetes
objects that are already normalized (among other things, lowercased)
with the expected test result in a CaC directory. In case the rule name
differs in case between the directory and the normalized name, we
wouldn't find the test result.