Fix Automatus --duplicate-templates
This option stopped working during the rework and now we are adding it
back, it basically will not save used templated test scenarios to the
list of used templated scenarios which will cause that the templated
scenarios will be used for each rule. This commit also removes unused
code because the method _rule_template_been_tested isn't called
anywhere.

Fixes: #9762