Ignore errors on semanage command when SELinux is not used
The remediation for the accounts_passwords_pam_faillock_dir rule also
ensures the proper SELinux File Context in the directory. The context is
applied using the semanage command, which depends on SELinux being used.
When the SELinux is not used, the command is expected to fail. However,
in the Ansible context, the entire Playbook execution is stopped. Since the
rule is actually asking for the dir and not for the SELinux context, it
is fair enough to make this Ansible task to be ignored when failed.
There is a specific rule for checking the SELinux File Context for
faillock logs: account_password_selinux_faillock_dir rule.