Remove unnecessary metric
Previously in the script there was a concept that partial coverage would
be the full coverage metric plus requirements in partial status. It
could be interpreted as requirements with "at least partial coverage".
However, this is not clear and is prone to misinterpretation. After
some analysis and feedback, it was concluded that "partial coverage"
should be restricted to requirements with partial status, without
summing the "full coverage" controls. This conclusion rendered the
"implicit metric" called "partial coverage" unusable, as the same data
is already reported in the metrics grouped by status.