Extract function recurse_or_substitute_or_do_nothing
This refactoring helps with passing a mutable object as a default
argument which is potentially dangerous - it is a global object, so if
the function adds anything to it during the execution, then subsequent
calls won't have a default of an empty list. For that reason, it is
common to have default values e.g. `None` and to have statements `s.a.
ignored_keys = ignored_keys or []` in the function.  This function is
split into a part that handles the outer loop and ignored keys, and
another part that handles the processing of individual values
`recurse_or_substitute_or_do_nothing`.