Improve the arguments for Ansible command module
It was used the awk command in a task and the delimiter character was
":". This has likely been worked around in the past, but not in the
most compatible way, causing the Playbook a fatal error on some systems
with different versions of Ansible. Relevant tasks have been improved
to be more robust and compatible.