Update sudo_require_reauthentication
Update OVAL checks to fail if more than one config
entry is found. Update and add tests accordingly.

Also update ansible and bash remediations to allow
whitespace surrounding 'timestamp_timeout'.

Update the sed commands on the bash remediation to use
the sed's named classes syntax instead of the traditional
regex syntax for character classes because sed was
not recognizing whitespace with the latter syntax.

Update rule yaml to mention the removal of duplicated
config entries on the fixtext.

Signed-off-by: Federico Ramirez <federico.r.ramirez@oracle.com>