Modify bash_pkg_conditional
This is a workaround for failed build. The problem is that the code in
the CPEItem.create_resolved_cpe_item_for_fact_ref method works with the
strings in Platform keys as if they are Python format strings and
substitutes values into the curly braces placeholders.  However, the
dpkg command uses curly braces literally in the command, which thean
causes the code to looks for a key to substitute into the format string,
and that raises a KeyError. We need to escape the curly braces in the
macro body. However, the macro is used also in other places than
"platform_package" template, where we don't want to have escaped curly
braces. The solution here is to have 2 variants, one with the escaped
curly braces in the string and other with the escaped curly braces.
This change is temporary. After all package-related platforms will use
the "platform_package" template, we will not need 2 versions of the
string. We will keep only the one that is with escaped curly braces.