Improve error handling when loading yaml stream
During tests with Ansible macros it was noticed a case where the Ansible
task names include "{{{ rule_title }}}" and an specific rule has special
symbols in its title. This was causing the load of the stream to fail
but the exception was not clear or easy to debug. So, it was created an
exception for this case giving some hints for the developer.