Improve Bash macro name and description
The name "bash_validate_authselect_custom_profile" was not so intuitive
about the variables defined there. On the other hand, the equivalent
macro in Ansible was much clearer. In order to make it more readable,
the macro was renamed to "bash_ensure_pam_variables_and_authselect_profile"
and a more complete description was included. It is now more readble and
more aligned to the equivalent in Ansible. In addition, a jinja comment
was included when a macro is called to set or modify a variable.