Fix Jinja condition in macro for pam_faillock
Recently the pam_faillock related macros were updated to include a
specific case for Ubuntu. In newer systems these changes were fine, but
in older systems, like RHEL7, the relevant code was not updated, causing
the remediation to fail due to files not found. Instead of duplicating
the existing conditional in two places, I move the testing logic to a
single place where the relevant files are defined. This should make it
simpler while fixing the issue.