Make benchmark inclusion optional
This commit introduces a new command line option `--include-benchmark`.
If this option is specified by the user, OVAL checks from rule
directories in the benchmark directory tree (eg. /linux_os/guide tree)
will be included in the generated OVAL. Adding these files is the
current behavior of the script so it doesn't add a new feature, instead
it allows users to turn this feature off by not providing this command
line option.

This commit also rewords and clarifies the help text for the
"ovaldirs" argument but it doesn't change the behavior of the "ovaldirs"
argument.