Rework OVAL rule to check time difference between now and last change date of shadow entry
There is strange issue I observed that if time_difference function is passed larger number of seconds than now(i.e. time in future)
It returns a very big integer instead of negative number so I had to monkeypatch the initial idea of the rule to check if the time difference is larger than 1000 years, then it means it was a negative result and we have a failure.

For example with the test code for fail scenario, i.e. putting last change date for tomorrow I get with time_difference the result 184467440737095,
while for dates in the past the result look ok.