Refactor and promote function to a class
The code has many issues that are triggering CodeClimate problems.
We tried to address them by simple refactoring, but then we realized
that more refactoring would make the code more readable and
understandable.

There are 2 branches of the code logic:
1. processing OVAL checks from benchmark directories (eg. /linux_os/guide)
2. processing OVAL checks from shared directories (eg. /shared/checks)

These 2 logic branches have similar code and it's possible to unify
them. However, I found the unification easier when I change the code
to a class.