Add Bash conditional for package version checking
For Debian packages, we will use the built-in function of `dpkg`.
For RPM packages, we won't use the `rpmdev-vercmp` command
because that's a part of the `rpmdevtools` which isn't installed by
default, but for most of the cases we can use the version comparison
algorithm from the `sort` command from the `coreutils` package.