Add a simpler example of the bash_pkg_conditional use
I expect the usage coconut>=4.5 to be the most common - if you have an
upstream version that is known to have a feature, then it is the upper
bound of all version numbers that may have it, as a feature may be
backported to an earlier version.  Therefore, various distributions
could specify their distribution-specific version specifiers if they are
after it, and a catch-all spec would look like a safe comparison against
the upstream version, not considering the release number and maybe even
not the minor number.