Add SLE15 based implementation of nftables_rules_permanent rule
The nftables_rules_permanent rule is based on the CIS requirement 3.5.2.10
There are some internal decisions made for the rule based on the facts that:
- Requirement assumes the system firewall is managed explicitely by nftables package and relevant kernel modules, i.e. no firewalld, no iptables installed
- SLE 15 nftables package comes with following number of configuration files:
  - /etc/nftables/arp-filter
  - /etc/nftables/bridge-filter
  - /etc/nftables/inet-filter
  - /etc/nftables/ipv4-filter
  - /etc/nftables/ipv4-mangle
  - /etc/nftables/ipv4-nat
  - /etc/nftables/ipv4-raw
  - /etc/nftables/ipv6-filter
  - /etc/nftables/ipv6-mangle
  - /etc/nftables/ipv6-nat
  - /etc/nftables/ipv6-raw
, where bridge, arp and inet(ipv4+ipv6), should cover L2+L3 layers of the IP stack, which are mainly addressed in other nftables requirements

So the design of the OVAL checks and Ansible and Bash remediation was made in a way that the rule will check that there is top-level configuration file,
as described by the requirement /etc/sysconfig/nftables, which will include per protocol family configuration files that come with the native package

The contents of the actual configuration as far as I uderstand the requirement should be left to the customer site policy.
The main idea of the rule as I tried to follow it was to make sure that there exists a configuration which to be applied on boot.

All of the changes made were SLE15 specific, since I based my logic on how the native nftables package for SUSE SLE15 looks like.