Include consideration about explicit parameters
This entry is not stating that all parameters should be explicit,
because this would also impact the readability in many cases.
Therefore, the intention of this entry is not to set a rule, but bring
awareness to the authors when writing Ansible tasks. Default values may
change along the time or may not be intuitive for others.