RHCOS4: Temporarily disable selinux_confinement_of_daemons to work around kubelet bug
kubelet changed its context to unconfined_service_t so this rule fails
and is impossible to remediate. We should push the kubelet developers to
provide a confined domain, but since there is no remediation, we need
to disable the rule in the meantime.

See https://issues.redhat.com/browse/OCPBUGS-6968 for more details.