Include RHEL reference for CIS firewalld zone rule
The 3.5.1.6 CIS requirement for RHEL7 and the 3.4.1.6 CIS requirement
for RHEL8 are manual. However, the set_firewalld_appropriate_zone rule
contains useful instructions to guide the users on how to check
interfaces and assign them to the appropriate zone. Including this rule
in the profile, makes it appear in the Guide.