Replace login_defs platform by package platform
The `login_defs` platform has slightly special meaning, the purpose
isn't to check for some `login_defs` package but the purpose is to check
for a package that provides `/etc/login.defs` file. However, there is
very little benefit in this logic. Instead, we can unify it and check
directly for the package because we know its name. To do that, we will
use the templated `package` platform. It is a different package on
different distributions, so we will as usually use the upstream project
name as the `package` platform key.