Fix automatus on Python 3.6
The os.path.dirname(__file__) on Python 3.6 seems to return an relative path
and on newer Pythons it returns an absolute path. This commit ensures that the
JINJA_MACROS_DIRECTORY is always an abssolute path. This is needed
because the AbsolutePathFileSystemLoader for Jinja will throw a
TemplateNotFound if the template path is not absolute