Increase robustness of OVAL check
The requirement for password expiration warning is valid only for
accounts with a password defined. Accounts without a password should be
ignored from the list and a filter was included for this.
It was also included a test to cover cases where there is no user with
a password defined in the system. Besides the improvements in tests,
this commit also align the OVAL to the project Style Guide.