Fix conditional in bash remediation
The previous condition was based on the result of systemctl
list-unit-files, which is the same regardless of the unit passed as
argument be present or not. The condition was updated to use grep
instead. A new test scenario was also included to make sure the rule
pass when there is not package installed and consequently no respective
socket unit.