Update pam_pwquality requirements in CIS
The CIS benchmark mentions that the try_first_pass option should be
present to satisfy the 5.5.1 requirement for RHEL8 and RHEL9 and the
5.4.1 requirement for RHEL7. However in these systems, the
pam_pwquality.so module, by default, is always the first module in the
password stack, making the "try_first_pass" useless and unnecessary
there. In addition, the "try_first_pass" is not present in the CIS
recommendation to remediate the system. In parallel to this update, it
was proposed updates for next CIS Benchmark versions.