Introduce variable for rule regex
The same regex was used twice. Considering it was a relatively long
regex, it was not easy to identify the similarities and also the
maintainability was compromissed when the regex needed updates.
It was introduced a variable for the regex and used this variable as
reference in the two places.