Ensure perm filter is considered in regex
This rule is about execution of privilleged commands. It should not care
about reading or writing in these files since there are other more
specific ways to check changes in these files. However, it is not
possible to say the rule without a filter for perm is wrong because it
would implicitly collect all permissions and not only execution. In
short, a rule without a perm filter would do much more than necessary
and it could lead to performance issues, but wouldn't be wrong in the
context of this rule. Based on this context, this commit includes the
perm filter in the check but makes it optional. The remdiation will be
updated to ensure the perm=x is actually present.