Improve OVAL check regarding eligible partitions
It was hard-coded in OVAL to ignore files if their file paths start
with some strings, which basically referred to pseudo filesystems or
filesystems that presumably should be ignored. This approach was not
robust and reliable enough. Instead, the partitions should be
dinamically filtered based on their mount options. This commit improves
the OVAL by dinamically detecting filtering out not relevant devices
and partitions mounted with noexec or nosuid mount options. This
restricts the scope of the probes to find suid and sgid files and
ultimately increases the check performance.