Include warning for NM Keyfiles in RHEL9
RHEL9 prefers and recommends to use NetworkManager keyfiles instead of
the ifcfg format for NIC configuration. New NICs will be automatically
configured in the NM keyfiles format by default. However, if the system
was upgraded from a RHEL8, the settings for existing NICs will remain
stored in ifcfg format. They should be migrated to the NM keyfiles
format by the Admins. It was included a warning in the
firewalld_sshd_port_enabled rule to create more awareness for admins.