Change macro bash_compare_version
This change is done to be able to use and test bash_compare_version
alone. It will no longer assume that the surrounding bash code defines
the `$real` and `$ver` variables, instead, they will be passed as
parameters, and that allows you to pass also specific strings
which can be handy for unit tests.