Remove check of the "new" AIDE database
The "new" database isn't needed for AIDE to work, only the
installed database matters, therefore, we shouldn't require
existence of the "new" database in our check. Sometimes,
the installed database is created by copying the "new" database,
but often it's created by moving the "new" database, which
is also suggested by the aide man page and the the "new" file
doesn't exist anymore.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2175684

This commit also adds a small regression test for the RHBZ.