Do not treat pull requests and issues differently
In Github API both pull requests and issues are treated as issues
objects. There are only some few differences in their attributes. While
the pyGithub have specific functions for both, the current function to
edit PRs does not accept milestone. This commit removes the script
option to update the milestone in PRs and unify this action in the
existing function to update milestone in Issues. The output messages
were also updated to inform the user.