Update CIS requirement status for umask requirement
The rule used to check and remediate the umask value in /etc/profile was
updated to also consider script files within /etc/profile.d, where is
also a valid place to define profile settings. This update was the
reason the respective CIS controls for RHEL were set as partial. They
are now automated.