remove @Base from RHEL kickstarts
The @base group is normally hidden (does not show up in
'dnf group list') and generally meant to be used as a base
for other, more visible, groups.

While the @base group itself is not deprecated, it likely
makes more sense to just rely on the default, which is
"Minimal install", a.k.a. @core, and let the users opt-into
additional packages.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>