Add control file for OCP 1.3.0
This commit uses a script still under development to generate a control
file for the OpenShift CIS 1.3.0 benchmark.

The control file isn't complete, but it does provide the boilerplate
control structure based on the benchmark.

This makes it easier for contributors to come through and fill in the
missing pieces, like rules and status.