Add initial component-rule mapping
This mapping has been generated by script `utils/assign_components.py`.
All rules from the `linux_os` benchmark directory have been assigned to
a component. Most of the rules are mapped to at least 1 component.  The
remaining rules are put into "unknown" component, but there is only 234
rules that are in the "unknown" component.  This mapping is wild. It
will have to be improved manually. Some rules will have to be add to
more components or moved to a different component.