fix test scenarios for file_(groupownership|owner|permissions) template
The test did not account for a case when a filepath is a directory and it does not exist on the system.
Also the test made use of is_directory parameter which is only available at runtime when building the template and it is not saved anywhere.
Instead of this parameter, scenarios now check if the path ends with a slash. This signifies the filepath is a directory as mentioned in the tempate documentation.