Rework jinja macro to use [[:blank:]] instead of \>
Strange enough usage of `\>` and `\` for escaping `.` and `*` produces unexpected results.
As in test string `*.* Something` does not match regex `^\*\.\*\>` but matches `^*.*\>`.
Since usage of the `\>` seems missin in other implmenetations(PCRE, Javascript etc), decided to change it to [[:blank:]],
and escape every special symbol like `*` or `.` with '\' for better readability.