create_srg_export: Add a command-line flag to prefer check and fix from the control over rules
The OCP4 STIG if a bit of an weird case. First, it mixes OCP4 and RHCOS
controls into a single stig and second, because of how the MachineConfig
remediations are not exactly user-friendly, we often lumped several
MachineConfigs from several rules into a single SRG row in the STIG
spreadsheet.

The current script presumes that every rule would emit a single row in
the sheet and that every rule's policy would be used exactly ones. The
most straightforward way of solving this is to just put the checks and
the fixes in the controls and then let the script optionally, with a new
command line flag that this patch adds, prefer those over policy files.

That way, we can have a single row even if the control specifies several
rules.

Additionally, because the current script presumes that rows whose check
and fix are read from the control files are either NA or IM and should
be separate in the resulting sheet, the script implements a rudimentary
deduplication of duplicate rows.

This behaviour is opt-in and would only be used by the OCP4 STIG for
the time being.