Refactor Ansible remediation
The find tasks were not working as expected due to a misinterpretation
of the "excludes" parameter. As a consequence some special dirs were
not ignored, hitting performance. Also, it was noticed a corner case
revelead by the test scenario where a NFS file system was targeting a
local folder. The corresponding local folder was not fixed due to the
way it was represented in Ansible Facts. All issues were solved in this
refactoring.