Fix `levels` key in general_profile script
The script we have for generating profiles was using the wrong key for
levels (it was using `level` instead of `levels`). This caused the
content build process to fail when you build the ocp content.

This commit updates the script to use the right key and also updates the
CIS OpenShift benchmark to use the right key.