Sanitize lines for clean YAML output when generating profiles
There is an issue with how pyyaml dumps lines, where if the line ends
with whitespace, styling won't matter.

  https://github.com/yaml/pyyaml/issues/121

This was affecting our ability to generate YAML output where audit
procedures (with many long lines) would get rendered as a single line
with escaped newlines. Instead, we want to render newlines as a block.

This commit parse each line of the data before it is dumped to make sure
trailing whitespace is removed as a workaround for this issue in pyyaml.