shared: templates: file_permissions: ansible: Improve handling of
directories in ansible remediation.
When we use the ansible file module to apply permissions only to
directories and subdirectories, this ends up applying to the files.
By using the find module we can prevent that and have a simple
remediation that works both for directories and file_regex.
This fixes #10687.