Include Ansible macro to create list of local paths
There are some rules which need to locate files and directories in a
system but should only consider local files and directories, ignoring
remote file systems and some known local directories with pseudo file
systems. Recently the Ansible approach was refactored in two rules in
order to make the remediation more robust and more efficient. A third
rule was discovered with the same issue so this macro was created to
make the remediation aligned among all relevant rules and to make the
same approach easier to be adopted by future rules.