Remove rules that cannot be applied during image build
Public Cloud images are built in a context that is different from a running
system. For example systemd is not running. As such systemctl start and
systemctl status will fail, see issue #10945 for details.

Remove firewalld rule outright; in the cloud other mechanisms exist for
handling connections to a running system and is generally the recommendation
not to run a firewall in the instance

Comment out the logrotate rule, we'd like to re-enable this after #10945 is
addressed

Comment out pam due to the way bash behaves. Requires a separate PR to address
the bash issue. No issue filed yet.