Remove CTest test "parse affected'
This test checks the presence and validity of the OVAL affected
element. Unfortunately, the test is extremely slow. Instead
of optimizing its speed, we need to realize that this OVAL
element is essential for the build and we have introduced
a check for this element at the build time which is implemented
in the `OVALBuilder._check_affected` method.