Add "test_all" subcommand
This commit adds a new subcommand `test_all` to the `fix_rules.py`
script.  The subcommand will execute all checks that are present in this
script on all rules in the product.  In other words, it works like
executing all currently existing subcommands in `--dry-run` mode at once.
We currently call the subcomands of this script in `--dry-run` mode
in our CTest test. This change will allow us to merge these tests
in a single test case.