Simplify display_login_attempts check
One one line of pam_lastlog.so without the silent option is enough.
In general the pam_lastlog.so is used to complement tools without the
native capability to show the "Last login" information but is used
without conditionals the information will be duplicated. This commit
simplify the OVAL in alignment to this expected behavior.