Remove test "validate-parse-platform"
This commit removes the CTest test `validate-parse-platform` and the
script `test_parse_platform.py` which performs this test.  This test
takes a long time - about 10 seconds, but its value is disputable. The
test tests the function `ssg.fixes.parse_platform()` by executing it on
each remediation file in the repository.  The tested function serves to
parse `# platform` expression in remediation headers. However, the
function isn't used during the build. During the build, we use
`ssg.build_remediations.split_remediation_content_and_metadata()` to
parse the `# platform` expression instead. We raise an exception there
during the build time if the platform can't be parsed. The tested
function `ssg.fixes.parse_platform()` is used only by the `mod_fixes.py`
utility.  An exception is raised in this script if the platform can't be
parsed.  As the integrity of the `# platform` expression in remediation
headers is ensured by a build time check and the tested code is used
only by the utility, I don't see a need for this time-intensive test in
the CI.