Update regex reference for new sshd_config line
In case the desired parameter is not yet present in sshd_config file, a
new line needs to be included. However, this line cannot be included at
the end of the file if there is already any "Match" conditional. In that
case, global parameters needs to be inserted before the "Match"
conditional. The former regex was causing issues in specific cases where
more than one line was matching the regex. This commit updates the regex
to ensure new lines are inserted at the beggining of the file.