Add a faster alternative for generating HTML guides
The new script `generate_guides.py` builds HTML guides for
each profile in the given SCAP source data stream and
also it builds a HTML index.

The script `generate_guides.py` applies the XSLT template
`/usr/share/openscap/xsl/xccdf-guide.xsl` that is shipped
in the `openscap` RPM package. That means that this tool
still depends on OpenSCAP, but doesn't call the `oscap` tool.

The advantage of the new solution is that the given SCAP source
data stream is parsed only once, which is faster that parsing
this big XML file for each profile.

Unfortunately, this new script needs the `lxml` library because
the `xml.etree` doesn't provide a XSLT processing feature.
Therefore, we will make the usage of the new script optional and
we will fall back to the existing solution if `lxml` isn't installed.
This way we will improve the CI and the user's machines experience
without having to break a build.