Make the output to stdout as default
The initial behaviour was expecting to save the metrics in a file by
using the write_to_textfile function from prometheus_client. There was
no specific function to redirect the output to stdout. However, it was
found another funcion which returns the metrics encoded in utf-8. This
function was used to make the stdout as default.