Introduce the controleval_metrics.py
This script depends on controleval.py but is intended to generate
metrics to be consumed by other tools. In this initial version, it is
generating metrics for Prometheus but other tools can be included in the
future, if necessary. These functions were initially included in
controleval.py but the drawnback was that prometheus_client module was
required by the script even when calling the stats sub-command and
making the module conditional would impact the readbility.