Fix broken profile resolution
First, rename the function `resolve_selections_with_rules` to
`apply_filter` to better express the purpose of this function.

Second, move the call of `apply_filter` to a correct place.
Previously, the function was checking if a rule ID is a rule that
exists and is applicable to the currently built product. It removed
the selections that selected the rules that don't exist. However,
this operation shadowed a check in `resolve` that served to raise
an exception if a rule isn't available. This exception
could never be raised because at the moment of the check guarding
the exception the selections already contained only existing rules.
This flaw is fixed by moving the `apply_filter` call after
the exception and removing the duplicate check.