Fail build if controls contain invalid rule IDs
We will raise an exception and terminate the build if a control
selects a rule that doesn't exist.

To do that, we need to get a list of all existing rules in the project.
Unfortunately, we can't reuse a list of the rules available in the
currently built product because control files can contain all rules from
all benchmarks from all products. Control files are product agnostic and
benchmark agnostic.