Refactored OVAL to make it simpler and more efficient
The previous logic was collecting all groups without members and then
comparing each group with a variable. In the report, when the rule
failed it was not clear the reason. The logic is broken in two tests
now. The first will search specifically for the desired group to make
sure the group exists. The second test will use the already collected
element and check the line against a regex to confirm the groups has no
members. The report is more transparent now and without unnecessary
information. The tests also reuse the same object.