Improve Ansible remediation
The remediation was not efficient by always removing the group even if
it was already compliant. Then it was again created. This is not
efficient and unnecessary changes were made. Currently there is no
option in user or group Ansible modules to remove members of a group. A
logic to achieve this would make the Playbook more complex than
necessary. Since this rule is only about local groups and users, we can
safely simplify the Playbook using lineinfile module. Ansible and Bash
are now aligned.