Review and update pcidss_4 requirement 2.2.7
Included configure_crypto_policy rule and removed rules related to
openssl and libreswan since these services are not required by the
policy. These rules were kept in related_rules to be reviewed next. We
should probably include platform on them so we can safely enable them in
this profile without forcing the installation of respective packages.