Review and update pcidss_4 requirement 4.2.1
Although libreswan or strongswan packages could be used to encrypt
PAN information during transmission, these packages are not mandatory.
Other encryption solutions can be used. Therefore, the two rules are
related to the requirement and were kept as reference but were note
selected to avoid installing unnecessary packages. Requirement is set to
supported as the system can use these technologies but the configuration
is too specific for each site in order to be automated.