Add profile aliases for OpenShift versioned profiles
We've implement profile versioning for some of the OpenShift and RHCOS
profiles. However, users that have ScanSettingBindings that reference
profiles like `ocp4-cis` can still be affected by rolling updates to
that profile. For example, when we implement support for CIS OpenShift
1.5.0, the `ocp4-cis` profile will automatically update to the rules for
that profile. This is how we've historically maintained profiles.

Now that we have a versioning mechanism and we're using it, we can give
users the ability to pin to a specific version of a profile.

This commit extends the profiles and names them according to their
version. When a user wants to pin to a specific version of a profile,
they can use `ocp4-cis-1-4` to run only the CIS OpenShift 1.4.0
rules, and they won't be affected by changes that implement the 1.5.0
profile when that is supported.

This change doesn't change the functionality of the operator or the
profiles, but gives users more flexibility for pinning to specific
profile versions.