Add a warning to rule service_rngd_enabled
This rule is used in STIG profile but the requirement isn't applicable
to RHEL 8.4 and newer.
See https://stigaview.com/products/rhel8/v1r12/RHEL-08-010471/

The applicability of the rule has already been fixed in
https://github.com/ComplianceAsCode/Content/commit/00513accd140621f730bd47297996a6d3b212fdf
This commit adds an warning explaining the reason for the
limited applicability.

Related to: https://issues.redhat.com/browse/RHEL-1819