Define prodtype for NTP service rules
These two rules are very similar but use a different name for the
service. It would be possible to unify these templated rules in the
future and deprecate one of them. For now, it is only included the
prodtype in both rules based on the current presence of the rule in
existing profiles and control files.