Introduce OVAL macro to map local mount points
This macro is expected to be used in many rules that need to check file
properties in the whole system. The nature of these rules is to consume
a lot of resources. This macro optimize the efficiency of these rules by
avoiding remote file systems as well as special file systems, like
/proc, /sys, etc.