Add new rule file_cron_allow_exists
Add new rule file_cron_allow_exists and add it to CIS profiles
because the CIS RHEL Benchmark requires the file /etc/cron.allow
to exist.

The other rules within the control that check the ownership and
permissions on /etc/cron.allow are passing if this file doesn't exist.
The file doesn't exist by default. To ensure the file exists, we add a
new rule that creates it.

Resolves: https://issues.redhat.com/browse/RHEL-1314