Represent references as lists internally
The reference values are stored in rule.yml files as strings
that are comma-separated lists of identifiers. We won't change
the rule.yml files stored in git because that would be a massive
change, but we will change the way how the references are stored
internally in build system and also in the resolved files.
This should simplify the code but won't have impact on built
SCAP data streams.