Fix Ansible compatibility with sysctl module
In older versions of Ansible, such as the version included with RHEL7,
the sysctl module was part of the ansible.builtin collection. But in
most recent versions it belongs to the ansible.posix collection. Since
explicity collection is not mandadotory, we can leave it implicit for
this particular case so each Ansible version can manage it properly.