CMP-2365: Fix check for rotating kubelet server certificates
The controller_rotate_kubelet_server_certs rule was only checking that
RotateKubeletServerCertificate was enabled by making sure it's enabled
explicitly. This doesn't work for clusters that are relying on the fact
that RotateKubeletServerCertificate is enabled by default.

This commit updates the check to invert the logic, so we're checking
that it's not explicitly disabled.