CMP-2372: Remove info override for virtual syscall rules
The STIG has controls that ensure virtual system calls are disabled,
because they pose a security risk. However, the rule was previously
overridden to always inform users, instead of actually checking the
functionality was disabled.

This commit removes the override so that we evaluate the kernel argument
appropriately. If a user still needs to disable this check (e.g., to
make it easier to run 32 bit applications) then they can still exclude
the rule using a TailoredProfile.