Add a new rule accounts_password_set_max_life_root
The new rule checks that maximum password age is set for the root user.
This rule is similar to rule `accounts_maximum_age_login_defs` but that
applies to all user accounts. The ANSSI password recommendation
(https://cyber.gouv.fr/publications/recommandations-relatives-lauthentification-multifacteur-et-aux-mots-de-passe)
recommends to not set the maximum age for common users, but it says that
for the privileged users the maximum age should be configured to
1 to 3 years. We concluded to agree on 1 year.