Improve ANSSI references
Use the URL that points directly to the PDF document with the ANSSI
Configuration Recommendations of a GNU/Linux System policy instead of
using the more generic link to ANSSI portal. This is similar to how we
link other benchmarks eg. PCIDSS. It allows users of HTML outputs to
directly navigate to the document to find the requirement text easily.
Also, this commit adds couple minor improvements, removal of additional
BP28 prefix, and tests updates.