Remove disa-stigs-uri from XSLT
The constants.xslt in the `eks` product is the only product-specific
constants XSLT file file that defines the variable `disa-stigs-uri`.
Other products don't define this variable. But, this product doesn't
have a STIG profile. So I think we can remove this variable. The
definition of this variable seems to be a copy paste problem, from other
products it was removed.